refactor(copilot): apply /simplify findings
- resolveTranscriptPath derives its root from the same Source() declaration
  the confiner guards, not a second derivation via sessionsDir(); they agree
  today only because Source() sets no DirByOS/DirFunc.
- copilotHome composes agentpaths.FromEnv+AbsRoot instead of re-deciding the
  override rule with a bare IsAbs, which silently dropped a relative value.
- the install-time version probe moves beside its only caller; hooks.go is
  the request path and said so while holding 50 lines of transcript scanning.
- the installer test uses hookjson's own HasOurHook instead of a hand-rolled
  walk that matched on the url's last segment and was weaker than production.
- name the cross-package hazard on the dispatch-only design guard.
